Move Python init to "where used/needed"
Driver is a regression on profile env vars, which were not being
applied because the Python VM was initialized before profile vars are
set.

This calls attention to any number of possible ordering issues that
may crop up.

Additionally, there's no reason to init Python on every command --
only some commands need Python.

`py::init` is idempotent so it's not a problem to call it repeatedly
thoughout the code.