Cleanup log dir resolution
- Use Path/PathBuffer up until we need to show something

- Remove use of NotFound error (and errors in general) to convey log
  status state

Gist is that we now always return a log dir, even if the log dir
doesn't exist. Inspect lazily creates the configured log dir so a
missing dir is not a problem we need to complain about.

Status shows missing dirs as dimmed as a hint that the log dir
doesn't (yet) exist.

See `tests/logs/log-dir.md` for handled scenarios.